feat: Support custom templates for commits + PRs
This commit introduces mustache templates for commit messages, and pull
request titles and bodies.

The design consists of templates with predefined properties. A custom
template can be provided.

Stubble has been chosen as the template renderer.

It is possible to define additional context for the templates, this also
includes delegates. These can be provided as the value of a "_delegates"
key in the context.