Fix allowed change to work per version (#1029)
* test: Showcase incorrect update behavior

Added test for `PackageUpdatesLookup` that demonstrates that updates are
incorrectly applied across different projects. `AllowedChange` needs to
be applied to each version independently, not across all projects as we
currently cannot guarantee that these different projects are loaded in
the same process at runtime.

* fix: Ensure AllowedChange works cross-project

In case a package is installed in multiple projects, allowed change will
only consider the highest installed version instead of looking at each
version independently.

This would result in an update being applied that is outside of the
allowed change range for projects for which the version of another
project with the highest version is considered outside of their allowed
change range.