feat: Support specifying threshold for open PRs (#1043)
Before, NuKeeper would only generate a single pull request as it would
reselect the same updateset every single time, unless some outside
factors would influence the result of its prioritization algorithm.

Now, you can specify the max number of open pull requests on a
per-repository basis. This is currently only supported for Azure
Devops/TFS, however only Azure Devops Server has been tested.

Since there are no straightforward APIs for figuring out the number of
open pull requests, especially when using a PAT with `Code (Read &
Write)`, heuristics are used to figure out the number of open pull
requests as best as possible.

First, the current user is fetched. If this fails, a user by the name of
`nukeeper@bot.com` will be fetched. If this fails, all pull requests in
the repository will be considered that have the label `nukeeper`. If
none of these things were possible, it's assumed that there are 0 open
pull requests.

When the parameter `--consolidate` is specified, the default value for
`--maxopenpullrequests` is 1, otherwise it is `maxpackageupdates`.